Refactoring ps-mcp pyproject and files
