+++ mcp/pyproject.toml
@@ -1,9 +1,5 @@
-[build-system]
-requires = ["setuptools>=42", "wheel"]
-build-backend = "setuptools.build_meta"
-
 [project]
-name = "psmcp"
+name = "ps-mcp"
 version = "0.85.1"
 description = "Adobe Photoshop automation using MCP"
 requires-python = ">=3.10"
@@ -19,33 +15,22 @@ dependencies = [
     "websocket-client>=1.8.0",
 ]
 
-[project.optional-dependencies]
-dev = [
-    "pytest>=7.0.0",
-    "black",
-    "isort",
-    "mypy",
+[project.scripts]
+ps-mcp = "ps_mcp:main"
+
+[build-system]
+requires = ["hatchling"]
+build-backend = "hatchling.build"
+
+[tool.uv]
+dev-dependencies = [
+    "freezegun>=1.5.1",
+    "pyright>=1.1.389",
+    "pytest>=8.3.3",
+    "ruff>=0.8.1",
 ]
 
 [tool.setuptools]
-py-modules = ["fonts", "logger", "psmcp", "socket_client"]
+py-modules = ["fonts", "logger", "socket_client"]
 
-[tool.black]
-line-length = 88
-target-version = ['py38']
-include = '\.pyi?$'
 
-[tool.isort]
-profile = "black"
-line_length = 88
-
-[tool.mypy]
-python_version = "3.10"
-warn_return_any = true
-warn_unused_configs = true
-disallow_untyped_defs = true
-disallow_incomplete_defs = true
-
-[tool.pytest.ini_options]
-testpaths = ["tests"]
-python_files = "test_*.py"
