+++ mcp/src/ps_mcp/__init__.py
@@ -1,0 +1,7 @@
+from ps_mcp.psmcp import mcp  
+
+def main():
+    mcp.run()
+
+if __name__ == "__main__":
+    main()
