+++ mcp/src/ps_mcp/__main__.py
@@ -1,0 +1,3 @@
+from ps_mcp import main   
+
+main()
